Disable dput progress indicator
It seems that in Travis CI scripts, the standard output is a tty.
As a result, upload progress indicator is enabled.
This is bad as it fills the Travis CI build log, then it is automatically
cancelled when it reaches 4 MB.

The trick is to pipe dput to "cat" to force its output to a pipe,
hence no progress is displayed.